+++ src/GRPC/CoordinatorRPCServer.cpp
@@ -23,7 +23,7 @@ CoordinatorRPCServer::CoordinatorRPCServer(CoordinatorEnginePtr coordinatorEngin
 
 Status CoordinatorRPCServer::RegisterNode(ServerContext*, const RegisterNodeRequest* request, RegisterNodeReply* reply) {
     NES_DEBUG("CoordinatorEngine::RegisterNode: request =" << request);
-    size_t id = coordinatorEngine->registerNode(request->address(), request->grpcport(), request->dataport(),
+    uint64_t id = coordinatorEngine->registerNode(request->address(), request->grpcport(), request->dataport(),
                                                 request->numberofslots(), request->nodeproperties(), (NodeType) request->type());
     if (id != 0) {
         NES_DEBUG("CoordinatorRPCServer::RegisterNode: success id=" << id);
